requirements.txt
absl-py==2.1.0
awkward==2.2.4
awkward-cpp==17
awkward0==0.15.5
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cycler==0.11.0
fonttools==4.38.0
google-auth==2.34.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
h5py==3.8.0
idna==3.8
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.3.2
kiwisolver==1.4.5
lz4==4.3.2
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
numexpr==2.8.6
numpy==1.21.6
oauthlib==3.2.2
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tables==3.7.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.13.1
torchinfo==1.8.0
torchvision==0.14.1
tqdm==4.66.5
typing-extensions==4.7.1
uproot==5.0.13
urllib3==2.0.7
vector==1.0.0
weaver-core==0.4.17
Werkzeug==2.2.3
xxhash==3.5.0
zipp==3.15.0
